d/rules: adjust GDCMTargets*.cmake
Some files are reaffected on the fly by multiarchitecture support,
pybuild, and possibly other scripts, after .cmake files are generated.
This causes various inconsistencies with the names of libraries,
between what is recorded in GDCMTargets*.cmake, and what is
effectively present in the file system.  This addresses the second
point of #989296, the first point being the missing dependencies to
libgdcm-dev.

Closes: #989296